backend/requirements.txt
# Viviz Bulk Sender - Django Backend

# Core
Django>=5.0,<6.0
djangorestframework>=3.14.0
django-filter>=23.0

# Database
psycopg2-binary>=2.9.9

# Authentication
djangorestframework-simplejwt>=5.3.0
django-rest-framework-simplejwt-cached-claims>=1.1.2
argon2-cffi>=23.1.0

# Multi-tenancy
django-tenants>=3.5.0

# Green API
requests>=2.31.0

# Task Queue
celery>=5.3.0
redis>=5.0.0
django-celery-beat>=2.5.0
django-celery-results>=2.5.0

# File Storage
django-storages>=1.14.0
boto3>=1.34.0

# Payment
stripe>=7.8.0

# Email
django-anymail>=10.2
sendgrid>=6.11.0

# Documentation
drf-spectacular>=0.26.0

# Utilities
python-dateutil>=2.8.2
pytz>=2024.1
uuid>=1.30

# Testing
pytest>=7.4.0
pytest-django>=4.7.0
pytest-cov>=4.1.0
factory-boy>=3.3.0

# Development
black>=23.11.0
flake8>=6.1.0
isort>=5.12.0
